requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bcrypt==3.2.0
PyPDF2>=3.0.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0
langchain>=0.1.0
tiktoken>=0.5.0
jieba>=0.42.1
scikit-learn>=1.0.2
numpy>=1.21.0